refactor: expose time domain intent wiring
